Fixing one more casting issue reported by the CI

+++ packages/dear-imgui/cpp/app/include/widget/plot_line.h
@@ -1,3 +1,5 @@
+#include <cstdint>
+
 #include <implot.h>
 #include "styled_widget.h"
 
@@ -59,7 +61,7 @@ public:
             return snprintf(buff, size,"0");
         }
 
-        return snprintf(buff, size, "%.*f", reinterpret_cast<int>(decimalPlaces), value);
+        return snprintf(buff, size, "%.*f", reinterpret_cast<intptr_t>(decimalPlaces), value);
     };
 
     bool HasCustomWidth() override;
